+++ examples/ex_quick.py
@@ -60,4 +60,4 @@ plt.plot(np.array(nes_hist['fitness']), label='NES')
 plt.xlabel('Generation')
 plt.ylabel('Fitness')
 plt.legend()
-plt.show()
+plt.close()   #change to plt.show() for inline plot